Rearrange code, rename functions, fix bookmarks
* Bookmarks were broken for types in dependencies files, because
  collectBookmarks was being run after filtering them out (whoops!)
* Make the executable as thin a wrapper as possible, mainly because this
  saves having to repeat myself in the cabal file.
* Move some code out of Main.hs into a new module Output.hs
* Rename functions for more sense and consistency (anything
  converting a RenderedThing into an OtherThing is now called
  thingAsOtherThing)
